requirements.txt
torch>=2.1.0
transformers>=4.39.0
datasets>=2.18.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0.1
sentencepiece
tiktoken
